refactor: move Domain and Application usecases to commands and queries folders
