+++ src/application/usecases/account/commands/index.ts
@@ -1,3 +1,2 @@
 export * from './db-add-account'
 export * from './db-authentication'
-export * from './db-get-account-id-by-token'
