+++ src/application/usecases/account/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './db-get-account-id-by-token'
