+++ src/domain/usecases/account/commands/index.ts
@@ -1,3 +1,2 @@
 export * from './add-account'
 export * from './authentication'
-export * from './get-account-id-by-token'
