+++ src/domain/usecases/account/queries/get-account-id-by-token.ts
