+++ src/main/factories/controllers/account/sign-up-controller-factory.ts
@@ -1,4 +1,4 @@
-import { AddAccountFactory, AuthenticationFactory } from '@/main/factories/usecases/account'
+import { AddAccountFactory, AuthenticationFactory } from '@/main/factories/usecases/account/commands'
 import { type Controller } from '@/presentation/protocols'
 import { SignUpController } from '@/presentation/controllers/account'
