+++ src/main/factories/middlewares/auth-middleware-factory.ts
@@ -1,4 +1,4 @@
-import { GetAccountIdByTokenFactory } from '@/main/factories/usecases/account'
+import { GetAccountIdByTokenFactory } from '@/main/factories/usecases/account/queries'
 import { type Middleware } from '@/presentation/protocols'
 import { AuthMiddleware } from '@/presentation/middlewares'
