+++ src/main/factories/usecases/account/commands/index.ts
@@ -1,3 +1,2 @@
 export * from './add-account-factory'
 export * from './authentication-factory'
-export * from './get-account-id-by-token-factory'
