+++ src/main/factories/usecases/account/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './get-account-id-by-token-factory'
