+++ src/presentation/middlewares/auth-middleware.ts
@@ -1,6 +1,6 @@
 import { type Middleware, type Response } from '@/presentation/protocols'
 import { HttpHelper } from '@/presentation/helpers'
-import { type GetAccountIdByToken } from '@/domain/usecases/account'
+import { type GetAccountIdByToken } from '@/domain/usecases/account/queries'
 import { InvalidCredentialsError, TokenError, AccessDeniedError } from '@/domain/errors/account'
 
 export class AuthMiddleware implements Middleware {
